// cloud-functions/functions/src/enums.ts
export enum PointType {
    Social,
    Activity,
    Creativity
}

export enum NotificationType {
    other = "other",
    comment = "comment",
    friendRequestReceived = "friendRequestReceived",
    friendRequestAccepted = "friendRequestAccepted",
    reportChallenge = "reportChallenge",
    reportComment = "reportComment",
    reportPost = "reportPost"
}

export enum ReportType {
    challenge = "challenge",
    comment = "comment",
    post = "post"
}

export enum ContentType{
    challenge = "challenge",
    comment = "comment",
    post = "post",
    goal = "goal",
    friendRequest = "friendRequest",

}